backend/requirements.txt
asgiref
Django
django-ckeditor
django-cors-headers
django-js-asset
djangorestframework
gunicorn
packaging
pillow
psycopg2-binary
dj-database-url
sqlparse
whitenoise
requests
